@@ -4,40 +4,58 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack-dev-server --progress --colors --content-base public"
+    "lint": "eslint app",
+    "start": "webpack-dev-server --progress --colors --content-base --port 8000"
+  },
+  "engines": {
+    "install-node": "7.6.0"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
     "babel-core": "^6.5.2",
+    "babel-eslint": "^7.1.0",
     "babel-loader": "^6.2.4",
+    "babel-plugin-import": "^1.1.0",
+    "babel-plugin-transform-builtin-extend": "^1.1.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-remove-strict-mode": "0.0.2",
+    "babel-plugin-transform-runtime": "^6.9.0",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-0": "^6.5.0",
+    "babel-runtime": "^6.11.6",
+    "classnames": "^2.2.5",
+    "config": "^1.25.1",
     "css-loader": "^0.23.1",
-    "express": "^4.13.4",
+    "eslint": "^3.6.0",
+    "eslint-config-airbnb": "^12.0.0",
+    "eslint-plugin-babel": "^3.0.0",
+    "eslint-plugin-import": "^1.6.1",
+    "eslint-plugin-jsx-a11y": "^2.0.1",
+    "eslint-plugin-markdown": "*",
+    "eslint-plugin-react": "^6.3.0",
+    "eslint-tinker": "^0.3.2",
     "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.8.5",
+    "file-loader": "^0.10.1",
     "html-webpack-plugin": "^2.9.0",
     "less": "^2.6.0",
     "less-loader": "^2.2.2",
+    "open-browser-webpack-plugin": "^0.0.4",
+    "pre-commit": "^1.2.2",
     "react-hot-loader": "^1.3.0",
+    "react-router-dom": "^4.0.0",
     "style-loader": "^0.13.0",
-    "url-loader": "^0.5.7",
+    "url-loader": "^0.5.8",
     "webpack": "^1.12.14",
     "webpack-dev-server": "^1.14.1"
   },
   "dependencies": {
-    "babel-preset-rax": "^0.0.9",
-    "bootstrap": "^3.3.6",
-    "classnames": "^2.2.5",
-    "core-decorators": "^0.15.0",
-    "element-loader": "^0.0.9",
-    "rax": "^0.0.9",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
-    "react-json-editor": "^0.2.0"
-  }
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "react-router": "^4.0.0"
+  },
+  "pre-commit": [
+    "lint"
+  ]
 }
